server/requirements.txt
fastapi
uvicorn
pydantic
mysql-connector-python
python-dotenv
PyJWT